Reorg folder structure for v2 behat bootstrap
